Split CQRS runtime behaviour tests (#118)
Separate hook/task-scheduling coverage from stream-pump runtime coverage so
each test module stays below the project line limit. Update the native wheel
snapshot to package both modules.
